1) quick refactor, moved code to separated folders, removed todo 2) added builder panel
