+++ src/pages/builder/style.css
@@ -2,3 +2,7 @@
   
 }
 
+#builder-page-wrapper {
+  position: relative;
+  height: 100%;
+}
