+++ src/pagesComponent/builderPageContent/components/blockCreationButton/style.scss
@@ -1,0 +1,10 @@
+.default-button-container {
+  width: 100%;
+  height: 100%;
+  background-color: transparent;
+  border: none;
+
+  &:hover {
+    
+  }
+}
