+++ src/pagesComponent/builderPageContent/components/builderPanelBar/index.tsx
@@ -1,0 +1,23 @@
+import { FC } from 'react'
+import cn from 'classnames'
+import './styles.scss'
+import BlockCreationButton from '../blockCreationButton'
+
+const BuilderPanelBar: FC = () => {
+  return (
+    <ul className={cn('builder-panel-bar__wrapper')}>
+      <li>
+        <BlockCreationButton />
+      </li>
+      <li></li>
+      <li></li>
+      <li></li>
+      <li></li>
+      <li></li>
+      <li></li>
+      <li></li>
+    </ul>
+  )
+}
+
+export default BuilderPanelBar;
