+++ src/pagesComponent/builderPageContent/components/builderPanelBar/styles.scss
@@ -1,0 +1,37 @@
+.builder-panel-bar__wrapper {
+  position: fixed;
+  right: calc(50% - 200px);
+  bottom: 10px;
+  width: 400px;
+  height: 50px;
+  display: flex;
+  background-color: rgba(24, 24, 24, 0.75);
+  border: 2px solid black;
+  list-style: none;
+
+  li {
+    width: 47px;
+    height: 47px;
+    border: 2px solid rgb(105, 105, 105);
+    cursor: pointer;
+
+    &:hover {
+      width: 44px;
+      height: 44px;
+      border: 4px solid rgb(196, 196, 196);
+    }
+  }
+
+  // TODO: Завести переменные для размеров экрана
+  @media screen and (max-width: 766px) {
+    right: calc(50% - 150px);
+    width: 320px;
+    height: 40px;
+
+    li {
+      width: 37px;
+      height: 37px;
+      cursor: pointer;
+    }
+  }
+}
